implement viewer page
