+++ app/ratel/src/features/spaces/pages/index/action_pages/mod.rs
@@ -1,0 +1,2 @@
+pub mod poll;
+pub use poll::*;
