+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/checkbox_choice/style.css
@@ -1,0 +1,1 @@
+/* Checkbox reuses option-multi CSS from multi_choice */
