+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/dropdown_choice/mod.rs
@@ -1,0 +1,2 @@
+mod component;
+pub use component::*;
