+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/dropdown_choice/style.css
@@ -1,0 +1,1 @@
+/* Dropdown reuses subjective-input CSS from subjective */
