+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/linear_scale/mod.rs
@@ -1,0 +1,2 @@
+mod component;
+pub use component::*;
