+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/linear_scale/style.css
@@ -1,0 +1,20 @@
+.scale-wrap { display: flex; flex-direction: column; gap: 16px; padding: 8px 0; }
+.scale-labels { display: flex; justify-content: space-between; }
+.scale-label { font-size: 12px; font-weight: 600; color: var(--dark, #8888a8) var(--light, #6b6b8a); max-width: 120px; }
+.scale-label--min { text-align: left; }
+.scale-label--max { text-align: right; }
+.scale-track { display: flex; gap: 8px; justify-content: center; flex-wrap: wrap; }
+.scale-point {
+  width: 52px; height: 52px; border-radius: 12px;
+  background: var(--dark, rgba(255,255,255,0.03)) var(--light, rgba(0,0,0,0.03));
+  border: 1px solid var(--dark, rgba(255,255,255,0.08)) var(--light, rgba(0,0,0,0.08));
+  display: flex; align-items: center; justify-content: center;
+  font-size: 16px; font-weight: 700;
+  color: var(--dark, #8888a8) var(--light, #6b6b8a);
+  cursor: pointer; transition: all 0.25s ease; padding: 0;
+}
+.scale-point:hover { background: rgba(129,140,248,0.06); border-color: rgba(129,140,248,0.2); transform: translateY(-2px); }
+.scale-point[data-selected="true"] { background: #818cf8; border-color: #818cf8; color: #fff; box-shadow: 0 4px 20px rgba(129,140,248,0.3); transform: translateY(-3px) scale(1.08); }
+.scale-point[data-in-range="true"] { background: rgba(129,140,248,0.08); border-color: rgba(129,140,248,0.25); color: #818cf8; }
+.scale-point[data-disabled="true"] { opacity: 0.5; cursor: not-allowed; pointer-events: none; }
+@media (max-width: 600px) { .scale-point { width: 44px; height: 44px; font-size: 14px; } }
