+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/mod.rs
@@ -1,0 +1,15 @@
+mod component;
+mod single_choice;
+mod multi_choice;
+mod subjective;
+mod linear_scale;
+mod checkbox_choice;
+mod dropdown_choice;
+
+pub use component::*;
+use single_choice::*;
+use multi_choice::*;
+use subjective::*;
+use linear_scale::*;
+use checkbox_choice::*;
+use dropdown_choice::*;
