+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/single_choice/style.css
@@ -1,0 +1,34 @@
+.options-single { display: flex; flex-direction: column; gap: 10px; }
+.option-single {
+  display: flex; align-items: center; gap: 14px; padding: 14px 18px; border-radius: 12px;
+  background: var(--dark, rgba(255,255,255,0.02)) var(--light, rgba(0,0,0,0.02));
+  border: 1px solid var(--dark, rgba(255,255,255,0.08)) var(--light, rgba(0,0,0,0.08));
+  cursor: pointer; transition: all 0.25s ease; position: relative; overflow: hidden;
+}
+.option-single:hover { background: rgba(129,140,248,0.04); border-color: rgba(129,140,248,0.2); }
+.option-single[data-selected="true"] {
+  background: var(--dark, rgba(99,102,241,0.15)) var(--light, rgba(99,102,241,0.08));
+  border-color: var(--dark, rgba(99,102,241,0.5)) var(--light, rgba(99,102,241,0.4));
+  box-shadow: 0 0 20px rgba(129,140,248,0.12);
+}
+.option-single__radio {
+  width: 20px; height: 20px; border-radius: 50%;
+  border: 2px solid var(--dark, rgba(255,255,255,0.08)) var(--light, rgba(0,0,0,0.08));
+  display: flex; align-items: center; justify-content: center; flex-shrink: 0; transition: all 0.25s;
+}
+.option-single[data-selected="true"] .option-single__radio { border-color: #818cf8; }
+.option-single__radio-dot {
+  width: 10px; height: 10px; border-radius: 50%; background: #818cf8;
+  transform: scale(0); transition: transform 0.25s cubic-bezier(0.34,1.56,0.64,1);
+}
+.option-single[data-selected="true"] .option-single__radio-dot { transform: scale(1); }
+.option-single__label { font-size: 15px; font-weight: 500; color: var(--dark, #f0f0f5) var(--light, #1a1a2e); flex: 1; }
+.option-single__letter {
+  font-size: 11px; font-weight: 700; letter-spacing: 0.06em;
+  color: var(--dark, #55556a) var(--light, #9999aa);
+  width: 24px; height: 24px; border-radius: 6px;
+  background: var(--dark, rgba(255,255,255,0.04)) var(--light, rgba(0,0,0,0.04));
+  display: flex; align-items: center; justify-content: center; flex-shrink: 0; transition: all 0.25s;
+}
+.option-single[data-selected="true"] .option-single__letter { background: rgba(99,102,241,0.08); color: #818cf8; }
+.option-single[data-disabled="true"] { opacity: 0.5; cursor: not-allowed; pointer-events: none; }
