+++ app/ratel/src/features/spaces/pages/index/action_pages/poll/subjective/style.css
@@ -1,0 +1,16 @@
+.subjective-wrap { display: flex; flex-direction: column; gap: 8px; }
+.subjective-input,
+.subjective-textarea {
+  width: 100%; padding: 16px; border-radius: 12px;
+  background: var(--dark, rgba(255,255,255,0.02)) var(--light, rgba(0,0,0,0.02));
+  border: 1px solid var(--dark, rgba(255,255,255,0.08)) var(--light, rgba(0,0,0,0.08));
+  color: var(--dark, #f0f0f5) var(--light, #1a1a2e);
+  font-family: 'Outfit', sans-serif; font-size: 15px; line-height: 1.7;
+  transition: border-color 0.25s; outline: none;
+}
+.subjective-textarea { min-height: 160px; resize: vertical; }
+.subjective-input::placeholder,
+.subjective-textarea::placeholder { color: var(--dark, #55556a) var(--light, #9999aa); }
+.subjective-input:focus,
+.subjective-textarea:focus { border-color: #818cf8; box-shadow: 0 0 0 3px rgba(129,140,248,0.12); }
+.subjective-counter { text-align: right; font-size: 12px; color: var(--dark, #55556a) var(--light, #9999aa); letter-spacing: 0.05em; }
